+++ src/interrupts/mod.rs
@@ -3,7 +3,6 @@ use core::sync::atomic::{compiler_fence, Ordering};
 use spin::Mutex;
 use crate::arch::x86_64::port_manager::{io_wait, Port};
 use crate::arch::x86_64::port_manager::ReadWriteStatus::{ReadWrite, WriteOnly};
-use crate::drivers::ps2::keyboard::PS2Keyboard;
 use crate::interrupts::interrupt_descriptor_table::*;
 use crate::interrupts::interrupt_service_routines::*;
 use crate::{println, print};
@@ -77,38 +76,38 @@ impl InterruptController {
     }
 
     fn map_handlers() {
-        IDT.set_entry(IdtVector::DivisionError, GateDescriptor::new(division_error_handler as u64));
-        IDT.set_entry(IdtVector::Debug, GateDescriptor::new(breakpoint_handler as u64));
-        IDT.set_entry(IdtVector::NonMaskableInterrupt, GateDescriptor::new(breakpoint_handler as u64));
-        IDT.set_entry(IdtVector::Breakpoint, GateDescriptor::new(breakpoint_handler as u64));
-        IDT.set_entry(IdtVector::Overflow, GateDescriptor::new(overflow_handler as u64));
-        IDT.set_entry(IdtVector::BoundRangeExceeded, GateDescriptor::new(bound_range_exceeded_handler as u64));
-        IDT.set_entry(IdtVector::InvalidOpcode, GateDescriptor::new(invalid_opcode_handler as u64));
-        IDT.set_entry(IdtVector::DeviceNotAvailable, GateDescriptor::new(device_not_available_handler as u64));
-        IDT.set_entry(IdtVector::DoubleFault, GateDescriptor::new(double_fault_handler as u64));
-        IDT.set_entry(IdtVector::InvalidTSS, GateDescriptor::new(invalid_tss_handler as u64));
-        IDT.set_entry(IdtVector::SegmentNotPresent, GateDescriptor::new(segment_not_present_handler as u64));
-        IDT.set_entry(IdtVector::StackSegmentFault, GateDescriptor::new(stack_segment_fault_handler as u64));
-        IDT.set_entry(IdtVector::GeneralProtectionFault, GateDescriptor::new(general_protection_fault_handler as u64));
-        IDT.set_entry(IdtVector::PageFault, GateDescriptor::new(page_fault_handler as u64));
-        IDT.set_entry(IdtVector::X87FloatingPointException, GateDescriptor::new(x87_floating_point_exception_handler as u64));
-        IDT.set_entry(IdtVector::AlignmentCheck, GateDescriptor::new(alignment_check_handler as u64));
-        IDT.set_entry(IdtVector::MachineCheck, GateDescriptor::new(machine_check_handler as u64));
-        IDT.set_entry(IdtVector::SIMDFloatingPointException, GateDescriptor::new(simd_floating_point_exception_handler as u64));
-        IDT.set_entry(IdtVector::VirtualizationException, GateDescriptor::new(virtualization_exception_handler as u64));
-        IDT.set_entry(IdtVector::ControlProtectionException, GateDescriptor::new(control_protection_exception_handler as u64));
-        IDT.set_entry(IdtVector::HypervisorInjectionException, GateDescriptor::new(hypervisor_injection_exception_handler as u64));
-        IDT.set_entry(IdtVector::VMMCommunicationException, GateDescriptor::new(vmm_communication_exception_handler as u64));
-        IDT.set_entry(IdtVector::SecurityException, GateDescriptor::new(security_exception_handler as u64));
+        IDT.set_entry(IdtVector::DivisionError, GateDescriptor::new(division_error_handler as usize));
+        IDT.set_entry(IdtVector::Debug, GateDescriptor::new(breakpoint_handler as usize));
+        IDT.set_entry(IdtVector::NonMaskableInterrupt, GateDescriptor::new(breakpoint_handler as usize));
+        IDT.set_entry(IdtVector::Breakpoint, GateDescriptor::new(breakpoint_handler as usize));
+        IDT.set_entry(IdtVector::Overflow, GateDescriptor::new(overflow_handler as usize));
+        IDT.set_entry(IdtVector::BoundRangeExceeded, GateDescriptor::new(bound_range_exceeded_handler as usize));
+        IDT.set_entry(IdtVector::InvalidOpcode, GateDescriptor::new(invalid_opcode_handler as usize));
+        IDT.set_entry(IdtVector::DeviceNotAvailable, GateDescriptor::new(device_not_available_handler as usize));
+        IDT.set_entry(IdtVector::DoubleFault, GateDescriptor::new(double_fault_handler as usize));
+        IDT.set_entry(IdtVector::InvalidTSS, GateDescriptor::new(invalid_tss_handler as usize));
+        IDT.set_entry(IdtVector::SegmentNotPresent, GateDescriptor::new(segment_not_present_handler as usize));
+        IDT.set_entry(IdtVector::StackSegmentFault, GateDescriptor::new(stack_segment_fault_handler as usize));
+        IDT.set_entry(IdtVector::GeneralProtectionFault, GateDescriptor::new(general_protection_fault_handler as usize));
+        IDT.set_entry(IdtVector::PageFault, GateDescriptor::new(page_fault_handler as usize));
+        IDT.set_entry(IdtVector::X87FloatingPointException, GateDescriptor::new(x87_floating_point_exception_handler as usize));
+        IDT.set_entry(IdtVector::AlignmentCheck, GateDescriptor::new(alignment_check_handler as usize));
+        IDT.set_entry(IdtVector::MachineCheck, GateDescriptor::new(machine_check_handler as usize));
+        IDT.set_entry(IdtVector::SIMDFloatingPointException, GateDescriptor::new(simd_floating_point_exception_handler as usize));
+        IDT.set_entry(IdtVector::VirtualizationException, GateDescriptor::new(virtualization_exception_handler as usize));
+        IDT.set_entry(IdtVector::ControlProtectionException, GateDescriptor::new(control_protection_exception_handler as usize));
+        IDT.set_entry(IdtVector::HypervisorInjectionException, GateDescriptor::new(hypervisor_injection_exception_handler as usize));
+        IDT.set_entry(IdtVector::VMMCommunicationException, GateDescriptor::new(vmm_communication_exception_handler as usize));
+        IDT.set_entry(IdtVector::SecurityException, GateDescriptor::new(security_exception_handler as usize));
 
-        IDT.set_irq_entry(0x20, GateDescriptor::new(irq0_handler as u64));
-        IDT.set_irq_entry(0x21, GateDescriptor::new(irq1_handler as u64));
-        IDT.set_irq_entry(0x22, GateDescriptor::new(irq2_handler as u64));
-        IDT.set_irq_entry(0x23, GateDescriptor::new(irq3_handler as u64));
-        IDT.set_irq_entry(0x24, GateDescriptor::new(irq4_handler as u64));
-        IDT.set_irq_entry(0x25, GateDescriptor::new(irq5_handler as u64));
-        IDT.set_irq_entry(0x26, GateDescriptor::new(irq6_handler as u64));
-        IDT.set_irq_entry(0x27, GateDescriptor::new(irq7_handler as u64));
+        IDT.set_irq_entry(0x20, GateDescriptor::new(irq0_handler as usize));
+        IDT.set_irq_entry(0x21, GateDescriptor::new(irq1_handler as usize));
+        IDT.set_irq_entry(0x22, GateDescriptor::new(irq2_handler as usize));
+        IDT.set_irq_entry(0x23, GateDescriptor::new(irq3_handler as usize));
+        IDT.set_irq_entry(0x24, GateDescriptor::new(irq4_handler as usize));
+        IDT.set_irq_entry(0x25, GateDescriptor::new(irq5_handler as usize));
+        IDT.set_irq_entry(0x26, GateDescriptor::new(irq6_handler as usize));
+        IDT.set_irq_entry(0x27, GateDescriptor::new(irq7_handler as usize));
     }
 
     fn remap_pic(offset_one: u8, offset_two: u8) {
